requirements.txt
# Core dependencies
pytest
playwright
pytest-playwright
rich
python-dotenv